chore: Move projects into src
