+++ CluedIn.Enricher.Web.sln
@@ -3,9 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 VisualStudioVersion = 17.0.31919.166
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{B2A52533-7FBF-41FE-A0AF-874FC54300B2}"
-	ProjectSection(SolutionItems) = preProject
-		src\Directory.Build.props = src\Directory.Build.props
-	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{7E2E897B-E664-4ECC-8AD8-814AAEAD8639}"
 	ProjectSection(SolutionItems) = preProject
@@ -44,11 +41,11 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Git", "Git", "{17EE3C8D-BAE
 		.gitignore = .gitignore
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExternalSearch.Providers.Web", "src\ExternalSearch.Providers.Web.csproj", "{E5832F12-6BB8-4ACA-AB23-F82888FF3971}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ExternalSearch.Web.Integration.Tests", "test\integration\ExternalSearch.Web.Integration.Tests\ExternalSearch.Web.Integration.Tests.csproj", "{DD79FB4E-FB22-41D7-8544-E2947EE6E99C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExternalSearch.Providers.Web.Provider", "ExternalSearch.Providers.Web.Provider\ExternalSearch.Providers.Web.Provider.csproj", "{663B5D87-5F55-40CD-9507-3C873A701E33}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExternalSearch.Providers.Web.Provider", "src\ExternalSearch.Providers.Web.Provider\ExternalSearch.Providers.Web.Provider.csproj", "{700063D5-D3DA-070C-C5AE-4F13F4100872}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ExternalSearch.Providers.Web", "src\ExternalSearch.Providers.Web\ExternalSearch.Providers.Web.csproj", "{DE111551-ACAC-CF3F-810E-D975CD5D6F59}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -56,18 +53,18 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E5832F12-6BB8-4ACA-AB23-F82888FF3971}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5832F12-6BB8-4ACA-AB23-F82888FF3971}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5832F12-6BB8-4ACA-AB23-F82888FF3971}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5832F12-6BB8-4ACA-AB23-F82888FF3971}.Release|Any CPU.Build.0 = Release|Any CPU
 		{DD79FB4E-FB22-41D7-8544-E2947EE6E99C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{DD79FB4E-FB22-41D7-8544-E2947EE6E99C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{DD79FB4E-FB22-41D7-8544-E2947EE6E99C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{DD79FB4E-FB22-41D7-8544-E2947EE6E99C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{663B5D87-5F55-40CD-9507-3C873A701E33}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{663B5D87-5F55-40CD-9507-3C873A701E33}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{663B5D87-5F55-40CD-9507-3C873A701E33}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{663B5D87-5F55-40CD-9507-3C873A701E33}.Release|Any CPU.Build.0 = Release|Any CPU
+		{700063D5-D3DA-070C-C5AE-4F13F4100872}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{700063D5-D3DA-070C-C5AE-4F13F4100872}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{700063D5-D3DA-070C-C5AE-4F13F4100872}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{700063D5-D3DA-070C-C5AE-4F13F4100872}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DE111551-ACAC-CF3F-810E-D975CD5D6F59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DE111551-ACAC-CF3F-810E-D975CD5D6F59}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DE111551-ACAC-CF3F-810E-D975CD5D6F59}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DE111551-ACAC-CF3F-810E-D975CD5D6F59}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -80,9 +77,7 @@ Global
 		{02641789-0D38-4DBB-9288-23592CF55D25} = {9BD54E81-D9C8-4419-82B8-8EE6CC7E2647}
 		{2C3AECCF-DCB5-4ED7-8B7B-8F373EDE3A5E} = {9BD54E81-D9C8-4419-82B8-8EE6CC7E2647}
 		{17EE3C8D-BAE4-481A-B9CC-1BE878E86124} = {9BD54E81-D9C8-4419-82B8-8EE6CC7E2647}
-		{E5832F12-6BB8-4ACA-AB23-F82888FF3971} = {B2A52533-7FBF-41FE-A0AF-874FC54300B2}
 		{DD79FB4E-FB22-41D7-8544-E2947EE6E99C} = {DBD1964B-DBD9-4C70-96CA-391C34B3D75F}
-		{663B5D87-5F55-40CD-9507-3C873A701E33} = {B2A52533-7FBF-41FE-A0AF-874FC54300B2}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6A61CB42-80C5-4A83-AC3E-DA41FC875525}
