+++ docs/4.6.0-release-notes.md
@@ -1,2 +1,5 @@
-### Features
+### Feature
 - Supports enricherv2
+
+### Chore
+- Move projects into src
