+++ src/ExternalSearch.Providers.Web/ExternalSearch.Providers.Web.csproj
