+++ src/ExternalSearch.Providers.Web/WebExternalSearchConstants.cs
