+++ src/ExternalSearch.Providers.Web/WebExternalSearchJobData.cs
