+++ src/ExternalSearch.Providers.Web/WebExternalSearchProvider.cs
